refactor(util) : unify browser util to state/browser
